@@ -1,16 +1,26 @@
-module github.com/mwitkow/grpc-proxy
+module github.com/glebkap/grpc-proxy
 
-go 1.14
+go 1.23.4
 
 require (
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/stretchr/testify v1.7.0
-	golang.org/x/net v0.0.0-20210331212208-0fccb6fa2b5c // indirect
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
-	golang.org/x/sys v0.0.0-20210331175145-43e1dd70ce54 // indirect
-	golang.org/x/text v0.3.6 // indirect
-	google.golang.org/genproto v0.0.0-20210401141331-865547bb08e2 // indirect
-	google.golang.org/grpc v1.36.1
-	google.golang.org/protobuf v1.26.0
-	honnef.co/go/tools v0.1.3
+	github.com/golang/protobuf v1.5.4
+	github.com/stretchr/testify v1.10.0
+	golang.org/x/sync v0.10.0
+	google.golang.org/grpc v1.69.4
+	google.golang.org/protobuf v1.36.3
+	honnef.co/go/tools v0.5.1
+)
+
+require (
+	github.com/BurntSushi/toml v1.4.1-0.20240526193622-a339e1f7089c // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/exp/typeparams v0.0.0-20231108232855-2478ac86f678 // indirect
+	golang.org/x/mod v0.17.0 // indirect
+	golang.org/x/net v0.30.0 // indirect
+	golang.org/x/sys v0.26.0 // indirect
+	golang.org/x/text v0.19.0 // indirect
+	golang.org/x/tools v0.21.1-0.20240531212143-b6235391adb3 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20241015192408-796eee8c2d53 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
